Read automation test counters as numbers, not integers
Luau integers in the Engine fork are a distinct 64-bit language type
with strict host-side reads, and the interpreter does not support
arithmetic on them (only native codegen does) - so script counters
incremented with += are correctly plain Lua numbers. The test harness
read those globals through the strict integer API, which yields 0 for
non-integer-tagged values; read them as numbers and cast host-side
instead.

Also bumps Submodules/Engine to the LinuxBuildFix commit reverting the
temporary double fallback in Zenith::ToInteger, restoring the intended
strict integer read semantics.

+++ Source/Game-Tests/Game-Tests/ScriptingAutomationTests.cpp
@@ -62,8 +62,10 @@ namespace
 
         i32 GetGlobalInteger(const char* name)
         {
+            // The counters are plain Lua numbers; Luau integers are a distinct type whose
+            // reads are strict, so read as number and cast host-side
             _zenith->GetGlobalKey(name);
-            const i32 value = _zenith->Get<i32>(-1);
+            const i32 value = static_cast<i32>(_zenith->Get<f64>(-1));
             _zenith->Pop();
             return value;
         }
